Synchronize dependencies with xcube 0.5.1 release
The packages and versions in the "run:" subsection of the "requirements"
section of the meta.yaml file now match those in the "Required" part of the
"dependencies" section of xcube's environment.yml, as it was at the 0.5.1
git release tag. Dependencies from the "Testing" section of environment.yml
are omitted from meta.yaml.